FRANKENSTEIN

+++ src/AccionSemantica/AS_Comentario.java
@@ -16,7 +16,7 @@ Agregar dígito al string
 
 public class AS_Comentario extends AccionSemantica {
     @Override
-    public int ejecutar(char car, Reader lector, StringBuilder token, TablaPalabrasReservadas PalabrasReservadas, Map<String, Simbolo> TablaDeSimbolos) {
+    public Short ejecutar(char car, Reader lector, StringBuilder token, TablaPalabrasReservadas PalabrasReservadas, Map<String, Simbolo> TablaDeSimbolos) {
     	AnalizadorLexico.token_actual.setLength(0); //VACIAMOS EL BUFFER YA QUE AL SER UN COM NO SE DEBE ALMACENAR 
         return TOKEN_ACTIVO;   
     };
