+++ src/AccionSemantica/AS_SIgChar.java
@@ -9,8 +9,8 @@ import Compilador.TablaPalabrasReservadas;
 
 public class AS_SIgChar extends AccionSemantica {
     @Override
-    public int ejecutar(char car, Reader lector, StringBuilder token, TablaPalabrasReservadas PalabrasReservadas, Map<String, Simbolo> TablaDeSimbolos) {
+    public Short ejecutar(char car, Reader lector, StringBuilder token, TablaPalabrasReservadas PalabrasReservadas, Map<String, Simbolo> TablaDeSimbolos) {
         //Esta accion semantica no hace nada mas que usarse para no cargar los caracteres q no se cargan al token.    
-        return TOKEN_ACTIVO;   
+    	return TOKEN_ACTIVO;   
     }; 
 }
